Remove parse_query from public/ui (#34232)
* Moved parse_query directive to agg_types
Moved fromQuery toQuery to query bar

* moved user input tests away from angular -> query_bar
