+++ src/legacy/ui/public/agg_types/buckets/filters.js
@@ -24,6 +24,7 @@ import { BucketAggType } from './_bucket_agg_type';
 import { createFilterFilters } from './create_filter/filters';
 import { decorateQuery, luceneStringToDsl } from '@kbn/es-query';
 import '../directives/click_focus';
+import '../directives/parse_query';
 import filtersTemplate from '../controls/filters.html';
 import { i18n } from '@kbn/i18n';
