+++ src/legacy/ui/public/query_bar/lib/from_user.ts
