+++ src/legacy/ui/public/query_bar/lib/index.ts
@@ -17,29 +17,6 @@
  * under the License.
  */
 
-import { toUser } from './lib/to_user';
-import { fromUser } from './lib/from_user';
-
-import { uiModules } from '../modules';
-uiModules
-  .get('kibana')
-  .directive('parseQuery', function () {
-
-    return {
-      restrict: 'A',
-      require: 'ngModel',
-      scope: {
-        'ngModel': '='
-      },
-      link: function ($scope, elem, attr, ngModel) {
-        const init = function () {
-          $scope.ngModel = fromUser($scope.ngModel);
-        };
-
-        ngModel.$parsers.push(fromUser);
-        ngModel.$formatters.push(toUser);
-
-        init();
-      }
-    };
-  });
+export * from './match_pairs';
+export * from './from_user';
+export * from './to_user';
