+++ src/legacy/ui/public/query_bar/lib/to_user.ts
